Move portal graphql code to src/graphql/portal
